need to determine why model output is NaN

+++ model/dataset.py
@@ -17,17 +17,17 @@ class HorseDataset(Dataset):
         row = self.df.iloc[idx]
 
         batch = {
-            "name":    torch.tensor(row["name"],   dtype=torch.long),
+            "name":    torch.tensor(row["name"],   dtype=torch.float32),
             "rating":  torch.tensor(row["rating"], dtype=torch.float32),
             "rawErg":  torch.tensor(row["rawErg"], dtype=torch.float32),
             "erg":     torch.tensor(row["erg"], dtype=torch.float32),
-            "age":     torch.tensor(row["age"], dtype=torch.long),
-            "sire":    torch.tensor(row["sire"],   dtype=torch.long),
+            "age":     torch.tensor(row["age"], dtype=torch.float32),
+            "sire":    torch.tensor(row["sire"],   dtype=torch.float32),
             "fee":     torch.tensor(row["fee"], dtype=torch.float32),
             "crop":    torch.tensor(row["crop"], dtype=torch.float32),
-            "dam":     torch.tensor(row["dam"],    dtype=torch.long),
+            "dam":     torch.tensor(row["dam"],    dtype=torch.float32),
             "ems3":    torch.tensor(row["ems3"], dtype=torch.float32),
-            "bmSire":  torch.tensor(row["bmSire"], dtype=torch.long),
+            "bmSire":  torch.tensor(row["bmSire"], dtype=torch.float32),
             "form":    torch.tensor(row["form"], dtype=torch.float32),
             "damForm": torch.tensor(row["damForm"], dtype=torch.float32),
             "numeric": torch.tensor(row[num_cols].values, dtype=torch.float32), # combines the hot encoded columns together
